Renamed command to `devbox generate devcontainer`
