+++ boxcli/root.go
@@ -28,7 +28,6 @@ func RootCmd() *cobra.Command {
 	command.AddCommand(AddCmd())
 	command.AddCommand(BuildCmd())
 	command.AddCommand(CloudCmd())
-	command.AddCommand(ContainerCmd())
 	command.AddCommand(GenerateCmd())
 	command.AddCommand(InfoCmd())
 	command.AddCommand(InitCmd())
